backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
torch
Pillow
